requirements.txt
absl-py==1.0.0
asgiref==3.5.0
astunparse==1.6.3
cached-property==1.5.2
cachetools==5.0.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
coreapi==2.3.3
coreschema==0.0.4
cryptography==36.0.1
cycler==0.11.0
defusedxml==0.7.1
Django==3.2.12
django-cors-headers==3.11.0
django-templated-mail==1.1.1
django-webpack-loader==1.5.0
djangorestframework==3.13.1
djangorestframework-simplejwt==4.8.0
flatbuffers==2.0
fonttools==4.32.0
gast==0.5.3
google-auth==2.6.4
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.44.0
h5py==3.6.0
idna==3.3
imbalanced-learn==0.7.0
imblearn==0.0
importlib-metadata==4.11.3
itypes==1.2.0
Jinja2==3.0.3
joblib==0.17.0
keras==2.8.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.2
libclang==13.0.0
Markdown==3.3.6
MarkupSafe==2.1.0
matplotlib==3.5.1
numpy==1.21.6
oauthlib==3.2.0
opt-einsum==3.3.0
packaging==21.3
pandas==1.3.5
patsy==0.5.2
Pillow==9.1.0
protobuf==3.20.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.3.0
pyparsing==3.0.8
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2021.3
reportlab==3.6.9
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
scikit-learn==0.23.2
scipy==1.7.3
seaborn==0.11.2
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.2.0
sqlparse==0.4.2
squarify==0.4.3
statsmodels==0.13.2
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.8.0
tensorflow-io-gcs-filesystem==0.24.0
termcolor==1.1.0
tf-estimator-nightly==2.8.0.dev2021122109
threadpoolctl==2.1.0
typing_extensions==4.1.1
uritemplate==4.1.1
urllib3==1.26.8
Werkzeug==2.1.1
wincertstore==0.2
wrapt==1.14.0
zipp==3.7.0
